+++ src/CameraGigePylon.h
@@ -227,7 +227,7 @@
             * @param value New exposure time value (us).
             * @return Success status to set new exposure time.
             */
-            bool setExposureTime(int exp);
+            bool setExposureTime(double exp);
 
             /**
             * Get device's gain value.
@@ -257,7 +257,7 @@
             *
             * @return Device's exposure time.
             */
-            int getExposureTime();
+            double getExposureTime();
 
     };
